rename utils to let language server figure out types
